main-page/add

+++ backblaze/views.py
@@ -17,7 +17,8 @@ def upload_image(request):
 
     image_validation(file_obj)
 
-    webp_image_name, webp_image_id, bucket_name = converter_to_webP(file_obj)
+    webp_image_name, webp_image_id, bucket_name, size = converter_to_webP(
+        file_obj)
 
     image_url = f'https://{bucket_name}.s3.us-east-005.backblazeb2.com/{webp_image_name}'
     file_model = FileModel(id=webp_image_id, name=webp_image_name,
@@ -25,7 +26,7 @@ def upload_image(request):
     file_model.save()
 
     return Response({
-                    'image_url': image_url, 'image_id': webp_image_id, 'image_name': webp_image_name
+                    'image_url': image_url, 'image_id': webp_image_id, 'image_name': webp_image_name, 'size': size
                     }, status=status.HTTP_200_OK)
 
 
@@ -49,9 +50,13 @@ def upload_document(request):
 def delete_file(request, file_id):
     try:
         file_model = FileModel.objects.get(id=file_id)
-        file_model.delete()
-        delete_file_from_backblaze(file_id)
-
-        return Response({'message': 'File deleted successfully'}, status=status.HTTP_200_OK)
+        FileNameAndId = delete_file_from_backblaze(file_id)
+        if FileNameAndId:
+            file_model.delete()
+            return Response({'message': 'File deleted successfully', 'file_name': FileNameAndId[0], 'file_id': FileNameAndId[1]}, status=status.HTTP_200_OK)
+        else:
+            return Response({'error': 'Failed to delete file from Backblaze'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
     except FileModel.DoesNotExist:
         return Response({'error': 'File not found'}, status=status.HTTP_404_NOT_FOUND)
+    except Exception as e:
+        return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
